+++ test/bash/UB/install-taskfile_01.sh
@@ -5,4 +5,5 @@ if [[ "$DEV_TEST_BASH_CONTAINER_ENVIRONMENT" != 'ON' ]]; then
   exit 1
 fi
 
-sudo /source/install-awscli
+export INST64_SYSTEM_WIDE='YES'
+sudo -E /source/install-awscli