+++ CHANGELOG.md
@@ -5,6 +5,44 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.3.0]
+
+### Fixed
+
+- BashLib64
+  - Wrong var name
+- YQ, Kind
+  - user-wide binary name
+
+### Changed
+
+- HadoLint
+  - upgraded installer skeleton
+
+### Added
+
+- Installers
+  - UV
+  - Terraform Docs
+- Trivy
+  - Multi Machine support
+- IstioCTL
+  - Multi Machine support
+- SHFMT
+  - Multi Machine support
+- Starship
+  - Multi Machine support
+- TFLint
+  - Multi Machine support
+- TFSec
+  - Multi Machine support
+- GitLeaks
+  - Multi Machine support
+- LocalStack
+  - Multi Machine support
+- Ruff
+  - Multi Machine support
+
 ## [5.2.0]
 
 ### Changed
@@ -104,7 +142,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Installer
   - The default is now to do non system-wide installation
 - All
-  - **breaking-change** Renamed parameter: INST64_<MODULE>_SYSTEM_WIDE -> INST64_SYSTEM_WIDE
+  - **breaking-change** Renamed parameter: INST64\_<MODULE>\_SYSTEM_WIDE -> INST64_SYSTEM_WIDE
 
 ## [3.41.2]